% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{asym.indices}
\alias{asym.indices}
\alias{limb.sym.index.1}
\alias{limb.sym.index.2}
\alias{limb.sym.index.3}
\alias{limb.sym.index.4}
\alias{bilat.strength.asym.1}
\alias{bilat.strength.asym.2}
\alias{bilat.asym.index.1}
\alias{bilat.asym.index.2}
\alias{asym.index.1}
\alias{sym.index}
\alias{sym.index.vect}
\alias{sym.angle}
\alias{percent.diff}
\alias{strength.asym}
\alias{asym.index.2}
\alias{btwn.limb.imbalance}
\title{Asymmetry Indices}
\usage{
asym.indices(r, l, dl, ndl)

limb.sym.index.1(dl, ndl)

limb.sym.index.2(dl, ndl)

limb.sym.index.3(l, r)

limb.sym.index.4(dl, ndl)

bilat.strength.asym.1(l, r)

bilat.strength.asym.2(l, r)

bilat.asym.index.1(dl, ndl)

bilat.asym.index.2(dl, ndl)

asym.index.1(dl, ndl)

sym.index(l, r)

sym.index.vect(l, r)

sym.angle(l, r)

percent.diff(l, r)

strength.asym(dl, ndl)

asym.index.2(l, r)

btwn.limb.imbalance(l, r)
}
\arguments{
\item{r}{Right limb value of length 1L}

\item{l}{Left limb value of length 1L}

\item{dl}{Dominant limb value of length 1L}

\item{ndl}{Non-dominant limb value of length 1L}
}
\value{
symmetry and asymmetry scores
}
\description{
Symmetry or asymmetry percentage as either scalar (always positive) or vector (indicates which leg is favoured).
}
\details{
There are 17 symmetry and asymmetry indices included within the package. Dependent on their format, the indices
can refer to the limbs as right (\code{r}) vs. left (\code{l}), dominant (\code{dl}) vs. non-dominant (\code{ndl}),
strong vs. weak, and high vs. low. The index functions take on either limbs identified as \code{l} and \code{r} or
\code{dl} and \code{ndl}. For indices that refer to the limbs as either strong vs. weak or high vs. low, the
function can assess which limb value is greater before computing the symmetry or asymmetry value.

The asymmetry index functions can be called upon explicitly via \code{interlimb:::}. Otherwise, users can call upon
\code{interlimb()} or \code{interlimb.wide()} to return one or all asymmetry index calculations, respectively.
}
\section{Functions}{
\itemize{
\item \code{limb.sym.index.1()}: Limb symmetry index 1

\item \code{limb.sym.index.2()}: Limb symmetry index 2

\item \code{limb.sym.index.3()}: Limb symmetry index 3

\item \code{limb.sym.index.4()}: Limb symmetry index 4

\item \code{bilat.strength.asym.1()}: Bilateral strength asymmetry 1

\item \code{bilat.strength.asym.2()}: Bilateral strength asymmetry 2

\item \code{bilat.asym.index.1()}: Bilateral asymmetry index 1

\item \code{bilat.asym.index.2()}: Bilateral asymmetry index 2

\item \code{asym.index.1()}: Bilateral strength asymmetry 1

\item \code{sym.index()}: Symmetry index

\item \code{sym.index.vect()}: Symmetry index, vectorized

\item \code{sym.angle()}: Symmetry angle

\item \code{percent.diff()}: Percent difference

\item \code{strength.asym()}: Strength asymmetry

\item \code{asym.index.2()}: Asymmetry index 2

\item \code{btwn.limb.imbalance()}: Between limb imbalance

}}
\note{
If users are interested in the algorithm, please see the accompanying publication or call upon the
asymmetry index function explicitly.
}
\examples{
# for reproducibility
set.seed(1)

# fabricate data
left.limb = 20
right.limb = 25

dominant.limb = 25
nondominant.limb = 20

pdiff.asym <- interlimb:::percent.diff(left.limb, right.limb)
lsi1.asym <- interlimb:::limb.sym.index.1(dominant.limb, nondominant.limb)
bsi1.asym <- interlimb:::bilat.strength.asym.1(left.limb, right.limb)

cat("pdiff: ", pdiff.asym,
"\nlsi1: ", lsi1.asym,
"\nbsi1: ", bsi1.asym)
}
