% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlimb.R
\name{interlimb}
\alias{interlimb}
\title{interlimb}
\usage{
interlimb(
  test.data,
  test.dates = "date",
  dominant.limb = "r",
  parameter,
  right.limb = "right",
  left.limb = "left",
  n.limbs = 2,
  best.n = c(1, 3, "all"),
  asym.algo = "auto",
  na.fill = TRUE
)
}
\arguments{
\item{test.data}{Original data set}

\item{test.dates}{Date variable name in quotation marks ("")}

\item{dominant.limb}{Individual's dominant limb as either \code{"r"} or \code{"l"}. Default is set to \code{"r"}}

\item{parameter}{The  parameter tested - does not need to be explicitly stated}

\item{right.limb}{Right limb variable name in quotation marks (""). Default set to \code{"right"}}

\item{left.limb}{Left limb variable name in quotation marks (""). Default set to \code{"left"}}

\item{n.limbs}{Number of limbs tested simultaneously as either \code{1} or \code{2}. Default set to \code{2}}

\item{best.n}{Takes on either an integer (i.e. 1, 2, 3) or \code{"all"} to assess either the best trial or mean of 2+ trials, respectively, per date.
If best.n is greater than the number of trials per date, all trials for the given date are selected. Default is set to \code{"all"}}

\item{asym.algo}{Asymmetry calculation/ algorithm to use, with default set to \code{"auto"}}

\item{na.fill}{Fill missing asymmetry values. Default set to \code{TRUE}}
}
\value{
A data.frame with the following variables: test.date, parameter, left.limb, right.limb,
dominant.limb, asymmetry.index, asymmetry, sd, cv, asym.change, rel.change, and rolling.cv
}
\description{
This is the primary function in the \code{interlimb} package. This function returns a data set that contains the
best or mean of the best n trials, asymmetry values, and reliability values grouped by date.
}
\details{
\code{interlimb::interlimb} is built with the understanding that users might be unfamiliar with R and tries to lower
the barrier of entry to maximize the user experience. The function takes on all necessary information and, in one
step, returns all symmetry and asymmetry information.

Users should note that some arguments must be input whereas others can be inferred by the function. For
example, variable names must be explicitly named and \code{n.limbs} stated. Conversely, if \code{n.limbs}
is not clarified, asym.algo will select the most appropriate calculations based on testing modality.

Since some asymmetry indices do not provide directionality (an indication of which limb is favored), the data.frame
that is returned to the user contains the variable \code{favored.limb}. In doing so, the user is always
provided limb symmetry or asymmetry directionality.
}
\note{
Please see the accompanying publication for extensive descriptions and examples concerning \code{interlimb} and
all of its functions.
}
\examples{
# for reproducibility
set.seed(1)

# fabricate data
asymmetry.tests <- data.frame(
    test.date = rep(LETTERS[1:5], each = 3),
    trial = rep(1:3, times = 5),
    left = rnorm(15, mean = 20, 1),
    right = rnorm(15, mean = 20, 1)
    )

# make the following assumptions:
# test was performed bilaterally (n.limbs = 2)
# dominant limb is right (dominant.limb = "r")
# the individual was tested during the countermovement jump

# the best trial is maintained for further analyses (best.n = 1)
bilateral.asym.max <- interlimb::interlimb(
    test.data = asymmetry.tests,
    test.dates = "test.date", # column name for the test dates
    dominant.limb = "r", # state dominant limb
    parameter = "jump height", # declare what was tested, can be set to NA
    right.limb = "right", # column name for the right limb
    left.limb = "left", # column name for the left limb
    n.limbs = 2, # was the test unilateral (1) or bilateral (2)
    best.n = 1, # take the best of the three trials per date
    asym.algo = "auto", # let the package decide which asymmetry index to use
    na.fill = TRUE # if data are missing, assumes that asymmetries remain constant week to week
    )

# test was performed unilaterally (n.limbs = 1)
# dominant limb is left (dominant.limb = "l")
# the individual was tested during the countermovement jump

# take all trials per day (n.best = "all)
unilateral.asym.mean <- interlimb::interlimb(
    test.data = asymmetry.tests,
    test.dates = "test.date",
    dominant.limb = "l",
    parameter = "jump height",
    right.limb = "right",
    left.limb = "left",
    n.limbs = 1,
    best.n = "all", # take the best of the three trials per date
    asym.algo = "auto", # let the package decide which asymmetry index to use
    na.fill = TRUE # if data are missing, assumes that asymmetries remain constant week to week
    )

bilateral.asym.max
unilateral.asym.mean
}
