% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlimb.helpers.R
\name{interlimb.wide}
\alias{interlimb.wide}
\alias{interlimb.long}
\title{Unabridged interlimb Data Frames}
\usage{
interlimb.wide(
  test.data,
  test.dates = "date",
  dominant.limb = "r",
  parameter,
  right.limb = "right",
  left.limb = "left"
)

interlimb.long(data.set)
}
\arguments{
\item{dominant.limb}{Dominant limb set as \code{"r"} or \code{"l"}}

\item{data.set}{A properly formatted data set (see details)}
}
\value{
Unabridged wide or long data.frame containing calculations for all asymmetry indices
}
\description{
\code{interlimb.wide()} and \code{interlimb.long()} return symmetry and asymmetry calculations for every asymmetry index that is
included in the package.
}
\details{
Since these functions are helper functions, the inputted data *must* be properly formatted. Therefore, the \code{data.set} argument
requires that the data.frame contain the testing date, parameter tested, and left and right limb assessments. The variables names must be:
\code{"test.date"}, \code{"parameter"}, \code{"left.limb"}, and \code{"right.limb"}. To set the variables names, use \code{colnames()}.

The `wide` and `long` functions are exported to help practitioners better decide which asymmetry index they want to choose for
the given testing modality. \code{interlimb.wide()} also provides the opportunity for researchers to explore relationships between asymmetry
indices.

If users are interested in exploring the best attempt versus mean of all attempts, they are encouraged to \code{aggregate()} the data by date
retain the best or all values, and calculate the mean. Examples are availale in either the \code{interlimb.R} source code or the
accompanying publication.

\code{interlimb.wide()} returns a data.frame where every asymmetry index is calculated as its own variable.

\code{interlimb.long()} returns a data.frame that has `combined` columns to create \code{asymmetry.index} and \code{asymmetry} variables.
}
\section{Functions}{
\itemize{
\item \code{interlimb.long()}: interlimb.long

}}
\examples{
# for reproducibility
set.seed(1)

# fabricate data
asymmetry.tests <- data.frame(
    test.date = rep(LETTERS[1:5], each = 3),
    trial = rep(1:3, times = 5),
    left = rnorm(15, mean = 20, 1),
    right = rnorm(15, mean = 20, 1)
    )

# calculate asymmetries for all tests
# assume dominant limb is the right limb

# wide data set
asymmetry.wide <- interlimb.wide(
    test.data = asymmetry.tests,
    test.dates = "test.date",
    dominant.limb = "r",
     parameter = "jump height",
     right.limb = "right",
     left.limb = "left")
head(asymmetry.wide)

# long data set
asymmetry.long <- interlimb.long(asymmetry.wide)
head(asymmetry.long)

# filter for asymmetry index
# subset(asymmetry.long, asymmetry.index == "percent.diff")
}
